requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
evidently
plotly
matplotlib
seaborn
requestspip 